Fix WalletConnect addChain on default chains
