+++ Assets/Resources/all_chains.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4fdff727c82be7446a43d7bd08f6f9c7
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
